requirements.txt
aiogram==3.22.0
aiogram-dialog==2.4.0
PyYAML==6.0.2
redis==6.4.0
openai==1.100.2
fluentogram==1.2.1
watchdog==6.0.0
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
httpx==0.28.1
git+https://github.com/molchanov-vs/my_tools.git@bb255b7